refactor: now use 'annotations Pinia store' and remove the 'Vuex annotations module'
