--- src/store/annotations/state.js
@@ -1,9 +1,0 @@
-export default function AnnotationState() {
-  return {
-    activeTab: '',
-    activeAnnotations: {},
-    annotations: null,
-    filteredAnnotations: [],
-    isLoading: false,
-  };
-}
