+++ src/store/index.js
@@ -1,11 +1,9 @@
 import { createStore } from 'vuex';
 
-import annotations from './annotations';
 import contents from './contents';
 
 export default createStore({
   modules: {
-    annotations,
     contents,
   },
 });
